+++ CHANGELOG.md
@@ -79,3 +79,4 @@
 * Add variance type arguments to `UT`, `UTU` and `TUT` schemes
 * Rename `U` to `TU`, `UU` to `TUV`, `UUU` to `TUVW` and put them into `Schemes` module
 * Define `Composition` typeclass and define its instances for `TU`, `TUV`, `TUVW`, `UT` and `UTU`
+* Define `Transformer` typeclass and define its instance for `Stateful` datatype